fai_m8_v1/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4

# Logging and utilities
structlog==24.1.0

# Pinecone - Vector database (Primary SERVICE)
pinecone-client==3.0.0

# Sentence Transformers - Local embeddings (no API key needed)
sentence-transformers==2.3.1
transformers==4.37.0
torch==2.1.2

# Machine Learning utilities
numpy==1.26.3
scikit-learn==1.4.0

# Jupyter support (for notebooks)
jupyter==1.0.0
jupyterlab==4.0.10
ipykernel==6.29.0
